+++ structuretimers/__init__.py
@@ -1,4 +1,4 @@
 default_app_config = "structuretimers.apps.StructureTimersConfig"
 
-__version__ = "1.1.0a1"
+__version__ = "1.1.0"
 __title__ = "Structure Timers"